Add primitive natural numbers (#41)
This introduces some primitives like unsigned numerals,
addition, (truncating) subtraction,  mutliplication,
division (throws on division by zero),
and a limited form of pattern-matching on natural numbers.

We use long arithmetic for the primitive naturals via the ZArith library.

The syntax for pattern-matching is also limited: it must always be
```
match ... with
| zero => ...
| succ ... => ...
end
```
i.e. one cannot swap the branches of the match-expression like so
```
match ... with
| succ ... => ...
| zero => ...
end
```

Recursion is of no concern for this change and will be added later.